install non-site-packages files in correct places
in particular, executables in "bin" or "Scripts" should now
be placed at the right location.

for some reason, the paths in the RECORD files do not contain the right
level of ".."s.
As far as I know, there's no place above site-packages but below toplevel
of the venv where there should be placed any files so we treat any
number of leading ".." s to mean "start from toplevel".

This also makes it impossible for a python package to install files above
the toplevel, but they shouldn't be allowed to do so anyway.