print err, but continue, if file in RECORD is missing
Pylint has non-existing files in their RECORD.
It's not evident from PEP 376 (or the PEP 627 draft), if that is valid.

It seems like the RECORD is mainly so safe removal is possible,
in which case non-existent files in it doesn't hurt, but it does
limit their usefulness for us
because we need a complete record of all files of the distribution.