change venv linking scheme further
Instead of creating just a single symlink to the venv in the central
store, create a folder and symlink in all the central venv's files and
folders in.

That makes it possible to store some additional metadata in the linked
virtpy, namely where it was created. This will allow us to better detect when a
virtpy was copied or moved.
There are quite a few potential error situations that a user might
cause inadvertently where cause and effect (breakage) are far apart
when moving virtpys around.
The metadata will allow virtpys to break in a more predictable and
understandable fashion.

To any python module that resolves symlinks and traverses the venv's
directory structure, this will still let them find all files in the
correct relative locations inside the venv.