remove obsolete workaround
With the old get-poetry.py installer, poetry created a batch script that
would not be picked up by Command::new("poetry").
Therefore we found a global python and called poetry via
`/path/to/python3 /path/to/poetry.py`.

With the new install-poetry.py installer, poetry creates an exe wrapper
for itself just like pip (and virtpy for that matter) does on windows.
This can be called via `Command` so manually passing the python path
around is no longer necessary.
Worse, calling the global python on poetry.py is broken now, because the
poetry script is now installed in a venv and does not manipulate the path
anymore to include its vendored dependencies before importing the main
function. It MUST be called with the python from the venv.