separate executable generation from global install
The functions for generating executables in a virtpy no longer need to
know about whether a package is globally installed.
The executables are symlinked to the binary directory after the virtpy
has been fully generated.

What executables belong to the requested package and not its dependencies
is loaded from the StoredDistribution.