pip_shim: call back the same virtpy that created it
The unit tests create a venv using virtpy which then calls back to virtpy
via the pip shim.
It would look for virtpy on the PATH and find the globally installed one.
That breaks the isolation of the test environment.
If no virtpy was globally installed, it just failed.
Thus, both in CI and locally, the install_uninstall test misbehaved.

This change stores the path to the creating executable in the venv's
metadata along with the project directory.